Move property getter fns from PPUCD Provider to UnicodeSet component
